FEAT: avatar data stored in local storage
avatar data gets fetched once and stored in local storage.
On the next load it will not be fetched but loaded from local storage.
This is to reduce fetching the server for the same data every page load.
Once the user logs out or changes his profile picture it will be fetched
again.